+++ android/src/main/java/com/applovin/reactnative/AppLovinMAXNativeAdView.java
@@ -403,22 +403,11 @@ public class AppLovinMAXNativeAdView
             nativeAdInfo.putDouble( "starRating", ad.getStarRating().doubleValue() );
         }
 
+        // The aspect ratio can be 0.0f when it is not provided by the network.
         float aspectRatio = ad.getMediaContentAspectRatio();
-        if ( !Float.isNaN( aspectRatio ) )
+        if ( aspectRatio > 0 )
         {
-            // The aspect ratio can be 0.0f when it is not provided by the network.
-            if ( Math.signum( aspectRatio ) == 0 )
-            {
-                nativeAdInfo.putDouble( "mediaContentAspectRatio", 1.0 );
-            }
-            else
-            {
-                nativeAdInfo.putDouble( "mediaContentAspectRatio", aspectRatio );
-            }
-        }
-        else
-        {
-            nativeAdInfo.putDouble( "mediaContentAspectRatio", 1.0 );
+            nativeAdInfo.putDouble( "mediaContentAspectRatio", aspectRatio );
         }
 
         nativeAdInfo.putBoolean( "isIconImageAvailable", ( ad.getIcon() != null ) );